NagrikAurSamvidhan-Backend/ChatBot/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
faiss-cpu
pypdf
langchain_community
langchain_groq
langchain_google_genai